Product Blacklist
Added a column to product table which indicates product status. If a new product is created we see if it fits into one of our categories. If it does we set status to approved and save. If not we search the blacklist keywords and mark as rejected. If we dont pick a category or blacklist we queue it to hit against image api.